fix conflict on load all fixtures
